Refactored API calls to use the `api` module.
