+++ src/components/Hamburger.tsx
@@ -8,7 +8,7 @@ type HamburgerProps = {
 const Hamburger: React.FC<HamburgerProps> = ({ open, onClick }) => (
   <button
     className={
-      `fixed top-4.5 left-4.5 z-[1002] w-11 h-11 bg-white border-2 border-[#f0f4fa] rounded-xl shadow-md flex flex-col justify-center items-center gap-1.5 cursor-pointer transition-shadow transition-colors duration-200
+      `fixed top-4.5 left-4.5 z-[1002] w-11 h-11 bg-white border-2 border-[#f0f4fa] rounded-xl shadow-md flex flex-col justify-center items-center gap-1.5 cursor-pointer transition-shadow duration-200
       ${open ? 'border-blue-500 shadow-lg' : ''}`
     }
     aria-label="Open menu"
@@ -22,4 +22,4 @@ const Hamburger: React.FC<HamburgerProps> = ({ open, onClick }) => (
   </button>
 );
 
-export default Hamburger; 
+export default Hamburger;
